Consolidate test files to improve organization
Merged 4 small test files into related larger test suites to reduce duplication
and improve test organization:

- test_none_regression.rs → option.rs (None-specific tests)
- slice_paths.rs → slices.rs (slice error path testing)
- test_simple_value_display.rs → enums.rs (error display tests)
- macro_expansion_debug.rs → basic.rs (simple macro test)

Benefits:
- Reduced test file count from 28 to 24 files
- Related tests now grouped in logical test suites
- Easier navigation and maintenance
- No functional changes - all tests still pass
- Maintains comprehensive test coverage
