+++ assert-struct/tests/slices.rs
@@ -515,3 +515,37 @@ error_message_test!(
 );
 
 error_message_test!("slices_errors/slice_pattern.rs", slice_pattern);
+
+// Slice path testing - error message path tracking for slice elements
+#[derive(Debug)]
+struct SlicePathData {
+    values: Vec<i32>,
+    scores: Vec<f64>,
+}
+
+#[test]
+fn test_slice_element_comparison_success() {
+    let data = SlicePathData {
+        values: vec![10, 20, 30],
+        scores: vec![85.5, 92.0, 78.3],
+    };
+
+    assert_struct!(data, SlicePathData {
+        values: [> 5, < 25, >= 30],
+        scores: [> 80.0, >= 90.0, < 80.0],
+    });
+}
+
+#[test]
+#[should_panic(expected = "data.values.[1]")]
+fn test_slice_element_comparison_failure() {
+    let data = SlicePathData {
+        values: vec![10, 5, 30],
+        scores: vec![85.5, 92.0, 78.3],
+    };
+
+    assert_struct!(data, SlicePathData {
+        values: [> 5, > 10, >= 30],  // Second element (5) fails > 10
+        scores: [> 80.0, >= 90.0, < 80.0],
+    });
+}
